+++ UltraStar Play/Assets/Common/Scene/SceneData/SingSceneData.cs
@@ -8,7 +8,7 @@ public class SingSceneData : SceneData
     public List<PlayerProfile> SelectedPlayerProfiles { get; set; } = new List<PlayerProfile>();
     public PlayerProfileToMicProfileMap PlayerProfileToMicProfileMap { get; set; } = new PlayerProfileToMicProfileMap();
     public Dictionary<PlayerProfile, string> PlayerProfileToVoiceNameMap { get; set; } = new Dictionary<PlayerProfile, string>();
-    public double PositionInSongInMillis { get; set; }
+    public float PositionInSongInMillis { get; set; }
     public bool IsRestart { get; set; }
     public int NextBeatToScore { get; set; }
     public Dictionary<PlayerProfile, PlayerScoreControllerData> PlayerProfileToScoreDataMap { get; set; } = new Dictionary<PlayerProfile, PlayerScoreControllerData>();
